2021.08.10 12:00 PM
modified the function that send user a page of each menu

The problem was in home.jsp.

	$("#seeMore").attr('href', '/listBoard?menu=' + menuName);

this statement was placed in the if loop which response exists.
so it didn't work when response doesn't exist (It means that there are no boardlist data like Spring we tested)

so the way how to resolve was...

1. I moved the statement out ot if loop.

2. set menu name variable to global

Plz check the script part in home.jsp.

Thank you😊